refactor: Table 패키징 작업
