+++ src/main/java/kitchenpos/dto/table/TableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.table;
 
 import static java.util.stream.Collectors.*;
