+++ src/main/java/kitchenpos/dto/table/TableGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.table;
 
 import kitchenpos.domain.table.OrderTableGroup;
